% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_substitutions}
\alias{apply_substitutions}
\title{Apply Template Substitutions}
\usage{
apply_substitutions(content, substitutions)
}
\arguments{
\item{content}{Character vector of content}

\item{substitutions}{Named list of substitutions}
}
\value{
Character vector with substitutions applied
}
\description{
Apply Template Substitutions
}
\keyword{internal}
