% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{available_templates}
\alias{available_templates}
\title{List Available Templates}
\usage{
available_templates(verbose = FALSE)
}
\arguments{
\item{verbose}{If TRUE, prints detailed descriptions. Default is FALSE.}
}
\value{
Character vector of available template names (invisibly if verbose).
}
\description{
Returns information about available project templates.
}
\examples{
available_templates()
available_templates(verbose = TRUE)
}
