% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.proj.R
\name{create.proj}
\alias{create.proj}
\title{Create project folder with preset or user given sub folders and a README file.}
\usage{
create.proj(path, parent, sub = c("code", "data", "deliverables",
  "documents"), doctype = "txt", setwd = FALSE)
}
\arguments{
\item{path}{A file path.}

\item{parent}{The project name. This becomes the top level folder for your project.}

\item{sub}{(optional) Sub folders to be created within your project folder.
Default folders created are "code","data","deliverables" and "documents".
User can supply any number of sub folders as a list c("folder1", "folder2").
User supplied folders will overwrite the default options.
For no subfolders, supply an empty list.}

\item{doctype}{(optional) Specify the type of document to be created as README.
Defoult is "txt". Any valid document type (i.e. "md", "doc") can be supplied.}

\item{setwd}{(optional) Sets the newly created project as the working directory. Default is FALSE}
}
\description{
Create project folder with preset or user given sub folders and a README file.
}
\examples{
create.proj(path = "C:/Users/username/Desktop", parent = "my_new_project")

#create a new folder with only 2 sub folders in it along with README.md. Then, set the new project as your working directory.
filepath <- "C:/Users/username/Desktop"
create.proj(path = filepath, parent = "Project_x", sub = c("Mulder", "Scully"), doctype = "md", setwd = TRUE)
}
\author{
Michael Golafshar
}
