% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{Create Default Config File}
\usage{
create_config(path = ".", global = FALSE)
}
\arguments{
\item{path}{Directory where config file should be created. Default is current directory.}

\item{global}{If TRUE, creates config in user's home directory (~/.kickstartR.yml).}
}
\value{
Invisibly returns the path to the created config file.
}
\description{
Creates a .kickstartR.yml configuration file with default settings.
This file can be customized to set team-wide or personal defaults.
}
\examples{
\dontrun{
# Create config in current directory
create_config()

# Create global config in home directory
create_config(global = TRUE)
}
}
