% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{create_gitignore}
\alias{create_gitignore}
\title{Create .gitignore}
\usage{
create_gitignore(path, template, include_renv, include_targets, extras = NULL)
}
\description{
Create .gitignore
}
\keyword{internal}
