% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{create_readme}
\alias{create_readme}
\title{Create README.md}
\usage{
create_readme(
  path,
  project_name,
  template,
  author,
  include_models,
  include_notebooks,
  create_rproj,
  include_renv,
  include_targets
)
}
\description{
Create README.md
}
\keyword{internal}
