% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{generate_license}
\alias{generate_license}
\title{Generate License File}
\usage{
generate_license(license, author = NULL, path)
}
\arguments{
\item{license}{License type. One of "MIT", "GPL-3", "GPL-2", "CC-BY-4.0", "CC0".}

\item{author}{Author name for the license. Defaults to system username.}

\item{path}{Directory where LICENSE file should be created.}
}
\value{
Invisibly returns the license text, or NULL if no license specified.
}
\description{
Creates a LICENSE file in the specified directory based on the license type.
}
\keyword{internal}
