% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Load all needed packages succinctly}
\usage{
packages(pkg, reinstall = FALSE, mirror)
}
\arguments{
\item{pkg}{A comma separated list of package names .}

\item{default}{(optional) Load default set of libraries. Default is set to TRUE
Default libraries include:
    Tidyverse - suite of data cleaning and manipulation tools
    Arsenal - functions for large-scale statistical summaries
    others????}

\item{group}{(optional) Set of additional libraries to load for specific tasks
i.e. viz, predict, descriptive, trees, ???}
}
\description{
Load all needed packages succinctly
}
\examples{
# load default packages
packages()

# load user defined package list
packages(rpart, pROC, rattle, rpart.plot, randomForest, caret) # Can we load without quotes???

#' # load group without default package
packages(group = "viz", default = FALSE)
}
\author{
Matthew Buras
}
