% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Set path}
\usage{
path(parent = FALSE, subfolder = FALSE, doc)
}
\arguments{
\item{parent}{(optional) The project name. This becomes the top level folder for your project.}

\item{sub}{(optional) Sub folders to within path to file.
If more than one subfolder, user can supply sub folders as a list c("folder1", "folder2").
Subfolders must listed be in the correct order from parent to file.}

\item{file}{Specify the file name}
}
\description{
Set path
}
\examples{
path(mydata.csv)
df <- read.csv(path(ISHLT, c(data, subdata, subsubdata), mydata.csv))
}
\author{
Matthew Buras, Michael Golafshar
}
