% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_template}
\alias{read_template}
\title{Read Template File}
\usage{
read_template(template_name)
}
\arguments{
\item{template_name}{Name of the template file}
}
\value{
Character vector of template contents
}
\description{
Read Template File
}
\keyword{internal}
