% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Load all needed packages succinctly}
\usage{

packages(pkg, reinstall = FALSE, mirror = "https://cloud.r-project.org/",
  quiet = TRUE)

}
\arguments{
\item{pkg}{A comma separated list of package names .}

\item{reinstall}{A logical indicating whether all of the packages provided should be installed or re-installed. Defalut is FALSE.}

\item{mirror}{A character string indicating the CRAN mirror to use to install packages. Default is mirror = "https://cloud.r-project.org/".}

\item{quiet}{A logical passed to install.packages for a quiet install. Default is TRUE.}
}
\description{
Load all needed packages succinctly
}
\examples{

Load collections by type of analysis?
Print message with packages installed
Can feed package list in c() form
Ex1: c("arsenal", "rpart", "survival") or as
a character string separated by a single space
Ex2: "arsenal rpart survival"
}
\author{
Matthew Buras
}
